+++ RTWWebServer/Controllers/AccountController.cs
@@ -1,7 +1,7 @@
 using Microsoft.AspNetCore.Identity.Data;
 using Microsoft.AspNetCore.Mvc;
 using NetworkDefinition.ErrorCode;
-using RTWWebServer.DTO.response;
+using RTWWebServer.DTOs.Response;
 using RTWWebServer.Services;
 
 namespace RTWWebServer.Controllers;
