--- RTWWebServer/Database/IDatabaseContextProvider.cs
@@ -1,6 +1,0 @@
-namespace RTWWebServer.Database;
-
-public interface IDatabaseContextProvider : IDisposable
-{
-    IDatabaseContext GetDatabaseContext(string databaseName);
-}
